refactor: Minor clean up to the code, with added comments.
This series of commits is work done by @dkihe for sending the station
number to the backend via an ajax post request.  This enables us pass
click events to the backend for scraping tide data.

This is a rough implementation and could be cleaned up a little more,
but this allows us to grab all the data we need without having an
intermediary database.